%
\name{symbs}
\alias{symbs}
\title{Transforme axis}
\usage{
symbs(symb.mean = NULL,
      symb.median = NULL,
      symb.five = NULL,
      symb.custom = NULL,
      minusplus.sign = FALSE,
      ...)
}
\arguments{
\item{symb.mean}{symb.mean}

\item{symb.median}{symb.median}
\item{symb.five}{symb.five}
\item{symb.custom}{symb.custom}
\item{minusplus.sign}{minusplus.sign}
\item{...}{other}

}
\description{
Letters
}
